rename order/index.js -> order.router.js, reformat code indents
